Add committee_meeting_people and committee_meeting_aliases tables
