+++ datapackage_pipelines_knesset/committees/processors/parse_committee_meeting_attendees.py
@@ -4,6 +4,8 @@ from datapackage_pipelines_knesset.common import object_storage
 from datapackage_pipelines_knesset.common import db
 import logging
 
+MAX_NAME_LEN = 70
+
 class ParseCommitteeMeetingAttendeesProcessor(BaseProcessor):
 
     def _process(self, datapackage, resources):
@@ -38,14 +40,14 @@ class ParseCommitteeMeetingAttendeesProcessor(BaseProcessor):
                     if key == "invitees":
                         yield {"committee_id":committee_id,
                                "meeting_id":meeting_id,
-                               "name":attendee["name"],
+                               "name":attendee["name"][:MAX_NAME_LEN] if len(attendee["name"]) > MAX_NAME_LEN else attendee["name"],
                                "role":"invitees",
                                "additional_information":attendee["role"] if "role" in attendee.keys() else ""}
 
                     else:
                         yield {"committee_id":committee_id,
                                "meeting_id":meeting_id,
-                               "name":attendee,
+                               "name":attendee[:MAX_NAME_LEN] if len(attendee) > MAX_NAME_LEN else attendee,
                                "role":key,
                                "additional_information":""}
